+++ src/streams.rs
@@ -1,4 +1,5 @@
 pub mod cluster;
+pub mod doc_registry;
 pub mod dynamic_attrs;
 pub mod jsite;
 pub mod psm_tables;